xml: Fix IOException in XmlPatternAnalysis when running the first time
Check first if state system file exists before calling delete on it.

Reviewed-on: https://git.eclipse.org/r/153726
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: CI Bot

+++ tmf/org.eclipse.tracecompass.tmf.analysis.xml.core/src/org/eclipse/tracecompass/internal/tmf/analysis/xml/core/pattern/stateprovider/XmlPatternAnalysis.java
@@ -150,11 +150,13 @@ public class XmlPatternAnalysis extends TmfAbstractAnalysisModule implements ITm
 
         if (!segmentStoreFile.exists()) {
             fStateSystemModule.cancel();
-            try {
-                Files.delete(stateSystemFile.toPath());
-            } catch (IOException e) {
-                // log it, failing to delete doesn't garantee failure
-                Activator.logWarning(e.getMessage(), e);
+            if (stateSystemFile.exists()) {
+                try {
+                    Files.delete(stateSystemFile.toPath());
+                } catch (IOException e) {
+                    // log it, failing to delete doesn't garantee failure
+                    Activator.logWarning(e.getMessage(), e);
+                }
             }
         }
